Implemented EventStatsModal
