+++ assets/js/pages/database/event_stats/index.js
@@ -1,5 +1,6 @@
-import eventStatsModal from './eventStatsModal.js';
 import boardActionsDropdown from '../../../elements/boardActionsDropdown.js';
+import eventAutocomplete from '../../../elements/eventAutocomplete.js';
+import eventStatsModal from '../../../elements/eventStatsModal.js';
 import gameStudyDropdown from '../../../elements/gameStudyDropdown.js';
 import historyButtons from '../../../elements/historyButtons.js';
 import ws from '../../../sanWs.js';
@@ -8,4 +9,4 @@ await ws.connect();
 
 localStorage.clear();
 
-eventStatsModal.modal.show();
+eventStatsModal.props.modal.show();
